Start functionality to upgrade product metadata
For stores with a large number of products we don't want to run the
update script on all products all at once. By using an admin page, we
can use ajax to update a smaller number of products at once while
showing a progressbar of the current state. This screen could also be
used for future updates that require similar functionality.